refactor(provider): split ProviderLoop god-file by concern (3577->525 LOC)
Split the 3,577-LOC ProviderLoop actor into focused same-module extension
files mirroring the existing ProviderLoop+SSEParser/+ErrorMapping pattern:
Serve, InferenceHandler, Preload, Prefetch, Testing, Trust, MemoryProtection,
IdleTimeout, Capacity, AutoUpdate, ModelLoading, Cancellation,
AttestationChallenge, LocalEndpoint.

Method bodies are byte-identical; only access control changed. Stored state
and methods reached across the split were widened private->internal (Swift
private is file-scoped); purely-local members stay private. Behavior unchanged.

Build green; swift test 1052/73 pass (unchanged from baseline).
